// src/types.ts
export interface Dictionary<TValue> {
  [id: string]: TValue;
}
export type AtLeast<T, K extends keyof T> = Partial<T> & Pick<T, K>;

export interface ServiceCall {
  service: string;
  entity_id?: string;
  service_data?: Dictionary<any>;
}

export interface Action {
  name?: string;
  service: string;
  service_data?: Dictionary<any>;
  icon?: string;
  variables?: VariableDictionary;
  supported_feature?: number;
}
// ...
export enum EVariableType {
  Level = 'LEVEL',
  List = 'LIST',
  Text = 'TEXT',
}

export interface Variable {
  name?: string;
  type: EVariableType;
  supported_feature?: number;
}

export interface ListVariableOption {
  value: string;
  icon?: string;
  name?: string;
}

export interface ListVariable extends Variable {
  options: ListVariableOption[];
}

export interface LevelVariable extends Variable {
  unit: string;
  min: number;
  max: number;
  step: number;
  scale_factor: number;
  optional: boolean;
}

export interface TextVariable extends Variable {
  multiline: boolean;
}

export type VariableDictionary = Dictionary<LevelVariable | ListVariable | TextVariable>;